Fixed getPath() not considering Y offset
The offset value of the original getPath() function is accidentally only a horizontal offset value; thus, if the current pos isn't on the same height as the target pos, the actual offside is bigger. Can cause the teleport not to work on servers because the real offset can exceed the max (vanilla) speed limit on servers even if the given offset value is low enough.