transformed workspace to regular design
if you are wondering what the hell I am going to do, I am going to revert the changes of the cross version and will keep all changes which have been done to b73. This version is actually going to be developed and released. Our main priority is to make a working workspace and version and keep important features like the support of scripts working. This is done to make sure we can release an updated version of the legacy build with support for microsoft login and all other changes required to guarantee a long lifetime.