LEGACY: Fixed incompatibility with Replay mod. (#744)
* transformed workspace to regular design

if you are wondering what the hell I am going to do, I am going to revert the changes of the cross version and will keep all changes which have been done to b73. This version is actually going to be developed and released. Our main priority is to make a working workspace and version and keep important features like the support of scripts working. This is done to make sure we can release an updated version of the legacy build with support for microsoft login and all other changes required to guarantee a long lifetime.

* Fixed LiquidBounce being incompatible with Replay mod.

* removed cross-version backend from code

* removed cross-version backend from code

* fixed artifact regex

* updated workflow

* added microsoft button to alt manager

this is not done but should work so far, so if you need to login into microsoft account use this.

* Revert "Fixed LiquidBounce being incompatible with Replay mod."

This reverts commit e31507364002d75e2c5fa0b2063d5be7e2888cdc.

* Fixed client being incompatible with Replay mod.

Co-authored-by: 1zuna <marco@ccbluex.net>
Co-authored-by: SenkJu <senkju4@gmail.com>