Reworked settings system
Now uses our API. It it still synced with our LiquidCloud GitHub, but it allows for more information to be passed to the client like type, last update, status of the settings and so on. We plan more features like uploading settings immediately to our API, which is one of the main purpose of this new system, but it needs more work on the API side.
Added 'Auto Settings' panel to the ClickGUI