Added options for KillAura's Packet AutoBlock mode. (#1282)
Let's explain these options:

The `releaseAutoBlock` option when enabled imitates a user who presses and then releases the right click button. When disabled, it instead imitates a user who just holds the right click button, resulting in full/constant blocking. That one is good for OP HvH situations.

The `ignoreTickRule` is based on one part of the game's blocking logic, which is when un-blocking. When it's time for the player to unblock the sword, it resets both the attack clicks and the useItem (right) clicks and skips to the next tick before you can attack and block again. When this option is enabled, it follows this logic, essentially simulating a legit auto block. When disabled, it ignores the tick delay and goes straight into the attack-block process.