Added a SafetyLines option to Scaffold. (#1367)
This option when enabled creates two lines that indicate most of the time safety when bridging using the GodBridge mode. For now it's best optimized for straight-bridging. These lines mean that you should step on them obviously, and there are 3, well 2 colors if the module rotates fast:

Green, which means you're good to go, especially when stepping on it.
Yellow, the rotation is nearly in a good aiming spot but there's a chance of falling if not corrected.
Red, which is danger. Rotation correction must be done otherwise you're falling.

The safety of the lines rely on the rotation difference between the target rotation and the current rotation. Oh and yeah the render code is not great, at least from me, like at all.