NEXTGEN: Fixed AutoArmor module's hotbar option not working and a invOpen option issue when on instant delay. (#1558)
Adapted to new ViaFabricPlus inventory open code which fixes AutoArmor constantly opening and closing inventories when on instant delay.

Hotbar option now will work using SilentHotbar to manage successful item interactions, as the previous one is not reliable at all.

Speaking of SilentHotbar, I removed the allowOverride check because it was messing with item interactions (not being on time when using fast delays). Plus what's the point of it anyway? Scaffold worked just fine even without it. If there's indeed issues removing it, then feel free to add it back and maybe improve it. To me it looks okay like this though.

Improved its delay logic. I also added StartDelay and CloseDelay options but they are commented for now, maybe we need them?