MCEF and Netty RestAPI/WS Implementation
Replaced ultralight implementation with MCEF. It uses CEF and is able to render web pages for the UI. This improves performance and also makes it easier to provide a known feature set for web developers.

For communication between the UI and the client, we will use our Netty RestAPI and WebSocket protocol. This protocol allows us to listen for events on the WebSocket while tasks are being executed via the RestAPI.

Clean-up and sorted

The following commits will extend the functionality of the RestAPI/WS to match our theme implementation from https://github.com/CCBlueX/LiquidBounce/pull/1442