refactor: BugUp -> AntiVoid (rewritten)
As Blink BugUp works on any anti-cheat, is the most reliable and was the default before, I have now removed all other modes and spent my time improving the Blink BugUp.
However, I have decided to rename the module to AntiVoid, as checking for fall damage was quite difficult in this case and also not very useful most of the time.
This module should be used for servers like Hypixel, Gomme, Gamster and Pika-Network. Everything you want...

It is not complete and has a few issues I have to deal with, but it is in a usable state.
Also, it will NOT stop you from falling when using Scaffold. Use the SafeWalk methods instead! It takes into account Sneak as well as Fly. It will only try to keep you from falling once.