feat: vomit module
Drops items from the inventory in a random order to make it look like you are vomiting.
If you are in creative mode, the module will drop random block items as quick as possible.
Can also take out items from containers if no slots are available to drop in inventory.