fix: CORS restrictions (#4168)
Restrict CORS to allow only localhost and 127.0.0.1 for development purposes.

CORS is not strictly required for our use case, as the REST API, WebSocket, and Static File Server are all hosted on the same host and port. However, to support development workflows—such as working on themes—we allow access to the Client REST API from `localhost` or `127.0.0.1`.

This change restricts access from third-party websites, ensuring that only local development environments can make cross-origin requests. Note that this change does not affect the `External Client Menu` feature, as it does not involve Cross-Origin requests when the host URL is accessed directly.