feat(mcef): accelerated paint support
This enables hardware-accelerated drawing by passing a shared texture from CEF to our MCEF renderer, rather than copying a byte buffer, which is much slower.

Currently, this only supports Windows on an NVIDIA GPU. AMD and Intel are both unsupported. For Intel, the following feature must be supported: https://github.com/IGCIT/Intel-GPU-Community-Issue-Tracker-IGCIT/issues/1143.
The situation with AMD is more problematic as support is only partial, and we have encountered crashes when tested on AMD.