fix(AntiVoid): repeating to rescue on failure
When AntiVoid failed to rescue us, it should stop trying. This is now the case. Also, the logic to determine if we are going to fall is drastically improved. Edge jumps are detecting without any issues, while when we would miss a block, we'd get rescued before we even fall.